Testing effect of adding energy state - and increased sampling to avoid action mask error.

+++ Services/base_service.py
@@ -13,7 +13,12 @@ tf.logging.set_verbosity(tf.logging.ERROR)
 
 def delete_nv_folder():
     location = "./../../.nv"
-    shutil.rmtree(location)
+    if os.path.isdir(location):
+        print("Correct dir, removing")
+        shutil.rmtree(location)
+        if os.path.isdir(location):
+            print("Failed...")
+
 
 class BaseService:
